refactor(dashboard-next): use barrel imports across the app
