+++ dashboard-next/src/components/ui/error-state.tsx
@@ -1,7 +1,7 @@
 import { AlertTriangle, type LucideIcon } from "lucide-react";
 import type { ReactNode } from "react";
-import { Button } from "@/components/ui/button";
 import { cn } from "@/lib/cn";
+import { Button } from "./button";
 
 interface ErrorStateProps {
   icon?: LucideIcon;
