+++ dashboard-next/src/components/ui/pagination.tsx
@@ -1,6 +1,6 @@
 import { ChevronLeft, ChevronRight } from "lucide-react";
-import { Button } from "@/components/ui/button";
 import { cn } from "@/lib/cn";
+import { Button } from "./button";
 
 interface PaginationProps {
   page: number;
