+++ dashboard-next/src/features/circuit-breakers/hooks.ts
@@ -1,5 +1,5 @@
 import { useQuery } from "@tanstack/react-query";
-import { useRefreshInterval } from "@/providers/refresh-interval-provider";
+import { useRefreshInterval } from "@/providers";
 import { fetchCircuitBreakers } from "./api";
 
 export function useCircuitBreakers() {
