+++ dashboard-next/src/features/dead-letters/hooks.ts
@@ -1,6 +1,6 @@
 import { keepPreviousData, useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
 import { toast } from "sonner";
-import { useRefreshInterval } from "@/providers/refresh-interval-provider";
+import { useRefreshInterval } from "@/providers";
 import { fetchDeadLetters, purgeDeadLetters, retryDeadLetter } from "./api";
 
 const KEY = {
