+++ dashboard-next/src/features/jobs/components/job-actions.tsx
@@ -2,8 +2,7 @@ import { useNavigate } from "@tanstack/react-router";
 import { Ban, Copy, RotateCcw } from "lucide-react";
 import { useState } from "react";
 import { toast } from "sonner";
-import { Button } from "@/components/ui/button";
-import { ConfirmDialog } from "@/components/ui/confirm-dialog";
+import { Button, ConfirmDialog } from "@/components/ui";
 import type { Job } from "@/lib/api-types";
 import { useCancelJob, useReplayJob } from "../hooks";
 import { canCancel, canReplay } from "../utils";
