+++ dashboard-next/src/features/jobs/components/job-errors-tab.tsx
@@ -1,7 +1,5 @@
 import { AlertOctagon } from "lucide-react";
-import { EmptyState } from "@/components/ui/empty-state";
-import { ErrorState } from "@/components/ui/error-state";
-import { Skeleton } from "@/components/ui/skeleton";
+import { EmptyState, ErrorState, Skeleton } from "@/components/ui";
 import type { JobError } from "@/lib/api-types";
 import { formatAbsolute } from "@/lib/time";
