+++ dashboard-next/src/features/jobs/components/job-logs-tab.tsx
@@ -1,7 +1,5 @@
 import { ScrollText } from "lucide-react";
-import { EmptyState } from "@/components/ui/empty-state";
-import { ErrorState } from "@/components/ui/error-state";
-import { Skeleton } from "@/components/ui/skeleton";
+import { EmptyState, ErrorState, Skeleton } from "@/components/ui";
 import type { TaskLog } from "@/lib/api-types";
 import { cn } from "@/lib/cn";
 import { formatAbsolute } from "@/lib/time";
