+++ dashboard-next/src/features/jobs/components/job-table.tsx
@@ -1,11 +1,7 @@
 import { useNavigate } from "@tanstack/react-router";
 import type { ColumnDef } from "@tanstack/react-table";
 import { useMemo } from "react";
-import { Badge } from "@/components/ui/badge";
-import { DataTable } from "@/components/ui/data-table";
-import { EmptyState } from "@/components/ui/empty-state";
-import { ErrorState } from "@/components/ui/error-state";
-import { Skeleton } from "@/components/ui/skeleton";
+import { Badge, DataTable, EmptyState, ErrorState, Skeleton } from "@/components/ui";
 import type { Job } from "@/lib/api-types";
 import { JOB_STATUS_LABEL, JOB_STATUS_TONE } from "@/lib/status";
 import { formatRelative } from "@/lib/time";
