+++ dashboard-next/src/features/resources/components/resources-table.tsx
@@ -1,11 +1,7 @@
 import type { ColumnDef } from "@tanstack/react-table";
 import { Activity } from "lucide-react";
 import { useMemo } from "react";
-import { Badge } from "@/components/ui/badge";
-import { DataTable } from "@/components/ui/data-table";
-import { EmptyState } from "@/components/ui/empty-state";
-import { ErrorState } from "@/components/ui/error-state";
-import { Skeleton } from "@/components/ui/skeleton";
+import { Badge, DataTable, EmptyState, ErrorState, Skeleton } from "@/components/ui";
 import type { ResourceStatus } from "@/lib/api-types";
 import { resourceTone } from "@/lib/status";
 import { formatDuration } from "@/lib/time";
