+++ dashboard-next/src/features/workers/components/workers-table.tsx
@@ -1,11 +1,7 @@
 import type { ColumnDef } from "@tanstack/react-table";
 import { Server } from "lucide-react";
 import { useMemo } from "react";
-import { Badge } from "@/components/ui/badge";
-import { DataTable } from "@/components/ui/data-table";
-import { EmptyState } from "@/components/ui/empty-state";
-import { ErrorState } from "@/components/ui/error-state";
-import { Skeleton } from "@/components/ui/skeleton";
+import { Badge, DataTable, EmptyState, ErrorState, Skeleton } from "@/components/ui";
 import type { Worker } from "@/lib/api-types";
 import { formatRelative } from "@/lib/time";
