+++ dashboard-next/src/routes/__root.tsx
@@ -1,7 +1,7 @@
 import { createRootRoute, Link, Outlet } from "@tanstack/react-router";
 import { AlertTriangle, ArrowLeft, Home } from "lucide-react";
 import { AppShell } from "@/components/layout";
-import { Button, buttonVariants } from "@/components/ui/button";
+import { Button, buttonVariants } from "@/components/ui";
 import { cn } from "@/lib/cn";
 
 export const Route = createRootRoute({
