+++ dashboard-next/src/routes/jobs/index.tsx
@@ -1,6 +1,6 @@
 import { createFileRoute } from "@tanstack/react-router";
 import { PageHeader } from "@/components/layout";
-import { Pagination } from "@/components/ui/pagination";
+import { Pagination } from "@/components/ui";
 import { JobFiltersBar, JobSearchBar, JobTable, useJobs } from "@/features/jobs";
 import type { JobFilters, JobListQuery } from "@/features/jobs/types";
 import { parseJobListSearch } from "@/features/jobs/utils";
